Rename Calculator.tsx to ProductionHelper.tsx to match component name
